backend/requirements.txt
Django==4.2.5
django-filter==23.2
djangorestframework==3.14.0
djoser==2.2.0
docker==6.1.3
flake8==6.1.0
gunicorn==21.2.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
reportlab==4.0.5